undo gmsh dependency cleanup
in the first commit of gmsh, sources that where not directly required for building the library where removed.
This makes upgrading harder because it's no easy to see what has been altered compared to the original
sources in order to successfully compile with Conan.

This is why the removed resources are now added back in.